requirements.txt
streamlit==1.40.1
google-generativeai==0.3.2
psycopg2-binary==2.9.11
sqlalchemy==2.0.23
pandas==2.2.3
plotly==5.18.0
numpy==2.1.1
pillow==11.0.0
pytesseract==0.3.10
pdf2image==1.16.3
python-dotenv==1.0.0
scipy==1.14.1
